third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/hidden-elements-expected.txt

This is a testharness.js-based test.
[FAIL] area should be hidden
  assert_equals: expected "none" but got "inline"
Harness: the test ran to completion.
